refactor(frontend): split task operation flows
