+++ packages/frontend/src/state/operations/tasks/task-mutation-runner.ts
@@ -1,0 +1,77 @@
+import { useQueryClient } from "@tanstack/react-query";
+import { useCallback } from "react";
+import { toast } from "sonner";
+import { errorMessage } from "@/lib/errors";
+import { refreshRepoTaskViewsFromQuery } from "@/state/queries/task-view-sync";
+import { requireActiveRepo } from "./task-operations-model";
+import type { TaskMutationRefreshStrategy } from "./task-operations-types";
+
+type RunTaskMutationOptions = {
+  refreshStrategy: TaskMutationRefreshStrategy;
+  run: (repoPath: string) => Promise<void>;
+  successTitle?: string;
+  successDescription: string;
+  failureTitle: string;
+};
+
+type UseTaskMutationRunnerArgs = {
+  activeRepoPath: string | null;
+};
+
+export type TaskMutationRunner = {
+  refreshTaskMutationViews: (
+    repoPath: string,
+    strategy: TaskMutationRefreshStrategy,
+  ) => Promise<void>;
+  runTaskMutation: (options: RunTaskMutationOptions) => Promise<void>;
+};
+
+export function useTaskMutationRunner({
+  activeRepoPath,
+}: UseTaskMutationRunnerArgs): TaskMutationRunner {
+  const queryClient = useQueryClient();
+
+  const refreshTaskMutationViews = useCallback(
+    async (repoPath: string, strategy: TaskMutationRefreshStrategy): Promise<void> => {
+      if (strategy.kind === "task") {
+        await refreshRepoTaskViewsFromQuery(queryClient, repoPath, {
+          forceFreshTaskList: true,
+          taskDocumentStrategy: "refresh",
+          taskIds: [strategy.taskId],
+        });
+        return;
+      }
+
+      if (strategy.kind === "remove-task") {
+        await refreshRepoTaskViewsFromQuery(queryClient, repoPath, {
+          forceFreshTaskList: true,
+          taskDocumentStrategy: "remove",
+          taskIds: strategy.taskIds,
+        });
+        return;
+      }
+
+      await refreshRepoTaskViewsFromQuery(queryClient, repoPath, { forceFreshTaskList: true });
+    },
+    [queryClient],
+  );
+
+  const runTaskMutation = useCallback(
+    async (options: RunTaskMutationOptions): Promise<void> => {
+      try {
+        const repoPath = requireActiveRepo(activeRepoPath);
+        await options.run(repoPath);
+        await refreshTaskMutationViews(repoPath, options.refreshStrategy);
+        if (options.successTitle) {
+          toast.success(options.successTitle, { description: options.successDescription });
+        }
+      } catch (error) {
+        toast.error(options.failureTitle, { description: errorMessage(error) });
+        throw error;
+      }
+    },
+    [activeRepoPath, refreshTaskMutationViews],
+  );
+
+  return { refreshTaskMutationViews, runTaskMutation };
+}
