+++ packages/frontend/src/state/operations/tasks/use-task-operations.ts
@@ -1,760 +1,66 @@
-import type {
-  BeadsCheck,
-  GitTargetBranch,
-  PullRequest,
-  RepoStoreHealth,
-  TaskCard,
-  TaskCreateInput,
-  TaskStatus,
-  TaskUpdatePatch,
-} from "@openducktor/contracts";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { toast } from "sonner";
-import { errorMessage } from "@/lib/errors";
-import type { TaskDataRefreshOptions, TaskRefreshOptions } from "@/state/app-state-contexts";
-import { documentQueryKeys } from "@/state/queries/documents";
-import { refreshRepoTaskViewsFromQuery } from "@/state/queries/task-view-sync";
-import { getBlockingRepoStoreHealth, summarizeTaskLoadError } from "@/state/tasks/task-load-errors";
-import type { ActiveWorkspace } from "@/types/state-slices";
-import { agentSessionQueryKeys } from "../../queries/agent-sessions";
-import { repoTaskDataQueryOptions } from "../../queries/tasks";
-import { settingsSnapshotQueryOptions } from "../../queries/workspace";
-import { host } from "../shared/host";
-import {
-  DEFERRED_BY_USER_REASON,
-  requireActiveRepo,
-  toNormalizedTitle,
-  toUpdateSuccessDescription,
-} from "./task-operations-model";
-
-type UseTaskOperationsArgs = {
-  activeWorkspace: ActiveWorkspace | null;
-  refreshBeadsCheckForRepo: (repoPath: string, force?: boolean) => Promise<BeadsCheck>;
-};
-
-type UseTaskOperationsResult = {
-  tasks: TaskCard[];
-  isForegroundLoadingTasks: boolean;
-  isRefreshingTasksInBackground: boolean;
-  isLoadingTasks: boolean;
-  detectingPullRequestTaskId: string | null;
-  linkingMergedPullRequestTaskId: string | null;
-  unlinkingPullRequestTaskId: string | null;
-  pendingMergedPullRequest: { taskId: string; pullRequest: PullRequest } | null;
-  setIsLoadingTasks: (value: boolean) => void;
-  clearTaskData: () => void;
-  refreshTaskData: (
-    repoPath: string,
-    taskIdOrIds?: string | string[],
-    options?: TaskDataRefreshOptions,
-  ) => Promise<void>;
-  refreshTasksWithOptions: (options?: TaskRefreshOptions) => Promise<void>;
-  refreshTasks: () => Promise<void>;
-  syncPullRequests: (taskId: string) => Promise<void>;
-  linkMergedPullRequest: () => Promise<void>;
-  cancelLinkMergedPullRequest: () => void;
-  unlinkPullRequest: (taskId: string) => Promise<void>;
-  createTask: (input: TaskCreateInput) => Promise<void>;
-  updateTask: (taskId: string, patch: TaskUpdatePatch) => Promise<void>;
-  setTaskTargetBranch: (taskId: string, targetBranch: GitTargetBranch) => Promise<void>;
-  deleteTask: (taskId: string, deleteSubtasks?: boolean) => Promise<void>;
-  resetTaskImplementation: (taskId: string) => Promise<void>;
-  resetTask: (taskId: string) => Promise<void>;
-  transitionTask: (taskId: string, status: TaskStatus, reason?: string) => Promise<void>;
-  deferTask: (taskId: string) => Promise<void>;
-  resumeDeferredTask: (taskId: string) => Promise<void>;
-  humanApproveTask: (taskId: string) => Promise<void>;
-  humanRequestChangesTask: (taskId: string, note?: string) => Promise<void>;
-};
-
-type TaskMutationRefreshStrategy =
-  | { kind: "repo" }
-  | { kind: "task"; taskId: string }
-  | { kind: "remove-task"; taskIds: string[] };
-
-const TASK_REFRESH_WARNING = "Pull request sync failed during task refresh";
-
-const collectTaskDeletionIds = (
-  tasks: TaskCard[],
-  taskId: string,
-  deleteSubtasks: boolean,
-): string[] => {
-  if (!deleteSubtasks) {
-    return [taskId];
-  }
-
-  const taskById = new Map(tasks.map((task) => [task.id, task]));
-  const collectedIds: string[] = [];
-  const pendingIds = [taskId];
-  const seenIds = new Set<string>();
-
-  while (pendingIds.length > 0) {
-    const currentId = pendingIds.shift();
-    if (!currentId || seenIds.has(currentId)) {
-      continue;
-    }
-
-    seenIds.add(currentId);
-    collectedIds.push(currentId);
-
-    for (const subtaskId of taskById.get(currentId)?.subtaskIds ?? []) {
-      if (!seenIds.has(subtaskId)) {
-        pendingIds.push(subtaskId);
-      }
-    }
-  }
-
-  return collectedIds;
-};
+import { useCallback } from "react";
+import { useTaskMutationRunner } from "./task-mutation-runner";
+import type { UseTaskOperationsArgs, UseTaskOperationsResult } from "./task-operations-types";
+import { useTaskMutationCommands } from "./use-task-mutation-commands";
+import { useTaskPullRequestOperations } from "./use-task-pull-request-operations";
+import { useTaskReadFlow } from "./use-task-read-flow";
+import { useTaskResetOperations } from "./use-task-reset-operations";
 
 export function useTaskOperations({
   activeWorkspace,
   refreshBeadsCheckForRepo,
 }: UseTaskOperationsArgs): UseTaskOperationsResult {
   const activeRepoPath = activeWorkspace?.repoPath ?? null;
-  const queryClient = useQueryClient();
-  const [isManualLoadingTasks, setIsManualLoadingTasks] = useState(false);
-  const manualRefreshTokenRef = useRef(0);
-  const inFlightTaskRefreshRef = useRef<{ repoPath: string; promise: Promise<void> } | null>(null);
-  const repoStoreHealthByRepoRef = useRef(new Map<string, RepoStoreHealth | null>());
-  const lastTaskRefreshToastRef = useRef<{
-    repoPath: string;
-    description: string;
-  } | null>(null);
-  const lastTaskLoadErrorToastRef = useRef<{
-    repoPath: string;
-    description: string;
-  } | null>(null);
-  const [detectingPullRequestTaskId, setDetectingPullRequestTaskId] = useState<string | null>(null);
-  const [linkingMergedPullRequestTaskId, setLinkingMergedPullRequestTaskId] = useState<
-    string | null
-  >(null);
-  const [unlinkingPullRequestTaskId, setUnlinkingPullRequestTaskId] = useState<string | null>(null);
-  const [pendingMergedPullRequest, setPendingMergedPullRequest] = useState<{
-    taskId: string;
-    pullRequest: PullRequest;
-  } | null>(null);
-  const currentWorkspaceRepoPathRef = useRef(activeRepoPath);
-  const settingsSnapshotQuery = useQuery(settingsSnapshotQueryOptions());
-  const settingsSnapshot = settingsSnapshotQuery.data ?? null;
-  const doneVisibleDays = settingsSnapshot?.kanban.doneVisibleDays ?? null;
-  const repoTaskDataQuery = useQuery({
-    ...repoTaskDataQueryOptions(activeRepoPath ?? "__disabled__", doneVisibleDays ?? -1),
-    enabled: activeRepoPath !== null && doneVisibleDays !== null,
+  const taskReadFlow = useTaskReadFlow({ activeRepoPath, refreshBeadsCheckForRepo });
+  const mutationRunner = useTaskMutationRunner({ activeRepoPath });
+  const mutationCommands = useTaskMutationCommands({
+    activeRepoPath,
+    tasks: taskReadFlow.tasks,
+    runTaskMutation: mutationRunner.runTaskMutation,
+  });
+  const resetOperations = useTaskResetOperations({
+    activeRepoPath,
+    refreshTaskData: taskReadFlow.refreshTaskData,
+  });
+  const pullRequestOperations = useTaskPullRequestOperations({
+    activeRepoPath,
+    refreshTaskData: taskReadFlow.refreshTaskData,
+    runTaskMutation: mutationRunner.runTaskMutation,
   });
 
-  useEffect(() => {
-    const previousActiveRepoPath = currentWorkspaceRepoPathRef.current;
-    currentWorkspaceRepoPathRef.current = activeRepoPath;
-    if (previousActiveRepoPath !== activeRepoPath) {
-      manualRefreshTokenRef.current += 1;
-      setIsManualLoadingTasks(false);
-      lastTaskRefreshToastRef.current = null;
-      lastTaskLoadErrorToastRef.current = null;
-      setDetectingPullRequestTaskId(null);
-      setLinkingMergedPullRequestTaskId(null);
-      setUnlinkingPullRequestTaskId(null);
-      setPendingMergedPullRequest(null);
-    }
-  }, [activeRepoPath]);
-
-  useEffect(() => {
-    const taskLoadError = settingsSnapshotQuery.isError
-      ? settingsSnapshotQuery.error
-      : repoTaskDataQuery.isError
-        ? repoTaskDataQuery.error
-        : null;
-
-    if (!taskLoadError || !activeRepoPath) {
-      if (!taskLoadError) {
-        lastTaskLoadErrorToastRef.current = null;
-      }
-      return;
-    }
-
-    const description = summarizeTaskLoadError({ error: taskLoadError });
-    const lastToast = lastTaskLoadErrorToastRef.current;
-    if (lastToast?.repoPath === activeRepoPath && lastToast.description === description) {
-      return;
-    }
-
-    lastTaskLoadErrorToastRef.current = { repoPath: activeRepoPath, description };
-    toast.error("Failed to load tasks", { description });
-  }, [
-    activeRepoPath,
-    repoTaskDataQuery.error,
-    repoTaskDataQuery.isError,
-    settingsSnapshotQuery.error,
-    settingsSnapshotQuery.isError,
-  ]);
-
-  const refreshTaskData = useCallback(
-    async (
-      repoPath: string,
-      taskIdOrIds?: string | string[],
-      options?: TaskDataRefreshOptions,
-    ): Promise<void> => {
-      const taskIds =
-        typeof taskIdOrIds === "string"
-          ? [taskIdOrIds]
-          : Array.isArray(taskIdOrIds)
-            ? taskIdOrIds
-            : null;
-      if (options?.source === "external-sync") {
-        await refreshRepoTaskViewsFromQuery(queryClient, repoPath, {
-          forceFreshTaskList: true,
-          ancillaryFailureMode: "best-effort",
-          ignorePrimaryCancellation: true,
-          refreshInactiveViews: false,
-          ...(taskIds
-            ? { taskDocumentStrategy: "invalidate", taskIds }
-            : { taskDocumentStrategy: "none" }),
-        });
-        return;
-      }
-
-      await refreshRepoTaskViewsFromQuery(
-        queryClient,
-        repoPath,
-        taskIds
-          ? { taskDocumentStrategy: "refresh", taskIds }
-          : {
-              forceFreshTaskList: options?.forceFreshTaskList ?? true,
-              taskDocumentStrategy: "none",
-            },
-      );
-    },
-    [queryClient],
-  );
-
-  const runRepoTaskRefresh = useCallback(
-    async (repoPath: string): Promise<void> => {
-      const beadsCheck = await refreshBeadsCheckForRepo(repoPath, false);
-      repoStoreHealthByRepoRef.current.set(repoPath, getBlockingRepoStoreHealth(beadsCheck));
-      await host.repoPullRequestSync(repoPath);
-      await refreshTaskData(repoPath);
-      try {
-        const refreshedBeadsCheck = await refreshBeadsCheckForRepo(repoPath, true);
-        repoStoreHealthByRepoRef.current.set(
-          repoPath,
-          getBlockingRepoStoreHealth(refreshedBeadsCheck),
-        );
-      } catch {
-        // Keep refresh semantics unchanged when the follow-up diagnostics check fails.
-      }
-    },
-    [refreshBeadsCheckForRepo, refreshTaskData],
-  );
-
-  const getRepoTaskRefreshPromise = useCallback(
-    (repoPath: string): { promise: Promise<void>; joinedExisting: boolean } => {
-      const inFlightRefresh = inFlightTaskRefreshRef.current;
-      if (inFlightRefresh && inFlightRefresh.repoPath === repoPath) {
-        return { promise: inFlightRefresh.promise, joinedExisting: true };
-      }
-
-      const promise = runRepoTaskRefresh(repoPath).finally(() => {
-        if (inFlightTaskRefreshRef.current?.promise === promise) {
-          inFlightTaskRefreshRef.current = null;
-        }
-      });
-      inFlightTaskRefreshRef.current = { repoPath, promise };
-      return { promise, joinedExisting: false };
-    },
-    [runRepoTaskRefresh],
-  );
-
-  const refreshTaskMutationViews = useCallback(
-    async (repoPath: string, strategy: TaskMutationRefreshStrategy): Promise<void> => {
-      if (strategy.kind === "task") {
-        await refreshRepoTaskViewsFromQuery(queryClient, repoPath, {
-          forceFreshTaskList: true,
-          taskDocumentStrategy: "refresh",
-          taskIds: [strategy.taskId],
-        });
-        return;
-      }
-
-      if (strategy.kind === "remove-task") {
-        await refreshRepoTaskViewsFromQuery(queryClient, repoPath, {
-          forceFreshTaskList: true,
-          taskDocumentStrategy: "remove",
-          taskIds: strategy.taskIds,
-        });
-        return;
-      }
-
-      await refreshRepoTaskViewsFromQuery(queryClient, repoPath, { forceFreshTaskList: true });
-    },
-    [queryClient],
-  );
-
-  const runTaskMutation = useCallback(
-    async (options: {
-      refreshStrategy: TaskMutationRefreshStrategy;
-      run: (repoPath: string) => Promise<void>;
-      successTitle?: string;
-      successDescription: string;
-      failureTitle: string;
-    }): Promise<void> => {
-      try {
-        const repoPath = requireActiveRepo(activeRepoPath);
-        await options.run(repoPath);
-        await refreshTaskMutationViews(repoPath, options.refreshStrategy);
-        if (options.successTitle) {
-          toast.success(options.successTitle, {
-            description: options.successDescription,
-          });
-        }
-      } catch (error) {
-        toast.error(options.failureTitle, {
-          description: errorMessage(error),
-        });
-        throw error;
-      }
-    },
-    [activeRepoPath, refreshTaskMutationViews],
-  );
-
-  const refreshTasksWithOptions = useCallback(
-    async (options?: TaskRefreshOptions): Promise<void> => {
-      if (!activeRepoPath) {
-        return;
-      }
-
-      const repoPath = activeRepoPath;
-      const trigger = options?.trigger ?? "manual";
-      let manualRefreshToken: number | null = null;
-      if (trigger === "manual") {
-        manualRefreshTokenRef.current += 1;
-        manualRefreshToken = manualRefreshTokenRef.current;
-        setIsManualLoadingTasks(true);
-      }
-
-      const { promise, joinedExisting } = getRepoTaskRefreshPromise(repoPath);
-
-      try {
-        await promise;
-        lastTaskRefreshToastRef.current = null;
-      } catch (error) {
-        const description = summarizeTaskLoadError({
-          error,
-          repoStoreHealth: repoStoreHealthByRepoRef.current.get(repoPath) ?? null,
-        });
-        if (!joinedExisting) {
-          console.warn(TASK_REFRESH_WARNING, {
-            repoPath,
-            trigger,
-            description,
-            error: errorMessage(error),
-          });
-        }
-
-        const lastToast = lastTaskRefreshToastRef.current;
-        const shouldDeduplicateToast =
-          lastToast?.repoPath === repoPath &&
-          lastToast.description === description &&
-          (trigger === "scheduled" || joinedExisting);
-
-        if (!shouldDeduplicateToast) {
-          lastTaskRefreshToastRef.current = { repoPath, description };
-          toast.error("Failed to refresh tasks", {
-            description,
-          });
-        }
-      } finally {
-        if (
-          trigger === "manual" &&
-          manualRefreshToken !== null &&
-          manualRefreshTokenRef.current === manualRefreshToken
-        ) {
-          setIsManualLoadingTasks(false);
-        }
-      }
-    },
-    [activeRepoPath, getRepoTaskRefreshPromise],
-  );
-
-  const refreshTasks = useCallback(async (): Promise<void> => {
-    await refreshTasksWithOptions({ trigger: "manual" });
-  }, [refreshTasksWithOptions]);
-
-  const syncPullRequests = useCallback(
-    async (taskId: string): Promise<void> => {
-      setDetectingPullRequestTaskId(taskId);
-      try {
-        const repoPath = requireActiveRepo(activeRepoPath);
-        const result = await host.taskPullRequestDetect(repoPath, taskId);
-        if (result.outcome === "linked") {
-          await refreshTaskData(repoPath, taskId);
-          toast.success("Pull request linked", {
-            description: `PR #${result.pullRequest.number}`,
-          });
-          return;
-        }
-        if (result.outcome === "merged") {
-          setPendingMergedPullRequest({
-            taskId,
-            pullRequest: result.pullRequest,
-          });
-          return;
-        }
-        toast.warning("No pull request found", {
-          description: `No open GitHub pull request found for ${result.sourceBranch}.`,
-        });
-      } catch (error) {
-        toast.error("Failed to detect pull request", {
-          description: errorMessage(error),
-        });
-      } finally {
-        setDetectingPullRequestTaskId((currentTaskId) =>
-          currentTaskId === taskId ? null : currentTaskId,
-        );
-      }
-    },
-    [activeRepoPath, refreshTaskData],
-  );
-
-  const cancelLinkMergedPullRequest = useCallback((): void => {
-    if (linkingMergedPullRequestTaskId != null) {
-      return;
-    }
-    setPendingMergedPullRequest(null);
-  }, [linkingMergedPullRequestTaskId]);
-
-  const linkMergedPullRequest = useCallback(async (): Promise<void> => {
-    if (!pendingMergedPullRequest) {
-      toast.error("Merged pull request state expired", {
-        description: "Re-run pull request detection and try again.",
-      });
-      return;
-    }
-
-    const { taskId, pullRequest } = pendingMergedPullRequest;
-    setLinkingMergedPullRequestTaskId(taskId);
-    try {
-      const repoPath = requireActiveRepo(activeRepoPath);
-      await host.taskPullRequestLinkMerged(repoPath, taskId, pullRequest);
-      setPendingMergedPullRequest((current) => (current?.taskId === taskId ? null : current));
-      await refreshTaskData(repoPath, taskId);
-      toast.success("Merged pull request linked", {
-        description: `PR #${pullRequest.number}; task moved to Done.`,
-      });
-    } catch (error) {
-      toast.error("Failed to link merged pull request", {
-        description: errorMessage(error),
-      });
-    } finally {
-      setLinkingMergedPullRequestTaskId((currentTaskId) =>
-        currentTaskId === taskId ? null : currentTaskId,
-      );
-    }
-  }, [activeRepoPath, pendingMergedPullRequest, refreshTaskData]);
-
-  const unlinkPullRequest = useCallback(
-    async (taskId: string): Promise<void> => {
-      setUnlinkingPullRequestTaskId(taskId);
-      try {
-        await runTaskMutation({
-          refreshStrategy: { kind: "task", taskId },
-          run: async (repoPath) => {
-            await host.taskPullRequestUnlink(repoPath, taskId);
-          },
-          successTitle: "Pull request unlinked",
-          successDescription: taskId,
-          failureTitle: "Failed to unlink pull request",
-        }).catch(() => {
-          // runTaskMutation already surfaced the actionable error to the user.
-        });
-      } finally {
-        setUnlinkingPullRequestTaskId((currentTaskId) =>
-          currentTaskId === taskId ? null : currentTaskId,
-        );
-      }
-    },
-    [runTaskMutation],
-  );
-
-  const createTask = useCallback(
-    async (input: TaskCreateInput): Promise<void> => {
-      requireActiveRepo(activeRepoPath);
-
-      const title = toNormalizedTitle(input.title);
-      if (!title) {
-        return;
-      }
-
-      await runTaskMutation({
-        refreshStrategy: { kind: "repo" },
-        run: async (repoPath) => {
-          await host.taskCreate(repoPath, {
-            ...input,
-            title,
-          });
-        },
-        successTitle: "Task created",
-        successDescription: title,
-        failureTitle: "Failed to create task",
-      });
-    },
-    [activeRepoPath, runTaskMutation],
-  );
-
-  const updateTask = useCallback(
-    async (taskId: string, patch: TaskUpdatePatch): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.taskUpdate(repoPath, taskId, patch);
-        },
-        successTitle: "Task updated",
-        successDescription: toUpdateSuccessDescription(taskId, patch),
-        failureTitle: "Failed to update task",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const setTaskTargetBranch = useCallback(
-    async (taskId: string, targetBranch: GitTargetBranch): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.taskUpdate(repoPath, taskId, { targetBranch });
-        },
-        successDescription: taskId,
-        failureTitle: "Failed to update task target branch",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const deleteTask = useCallback(
-    async (taskId: string, deleteSubtasks = false): Promise<void> => {
-      const taskIdsToRemove = collectTaskDeletionIds(
-        repoTaskDataQuery.data?.tasks ?? [],
-        taskId,
-        deleteSubtasks,
-      );
-      await runTaskMutation({
-        refreshStrategy: { kind: "remove-task", taskIds: taskIdsToRemove },
-        run: async (repoPath) => {
-          await host.taskDelete(repoPath, taskId, deleteSubtasks);
-        },
-        successTitle: "Task deleted",
-        successDescription: taskId,
-        failureTitle: "Failed to delete task",
-      });
-    },
-    [repoTaskDataQuery.data?.tasks, runTaskMutation],
-  );
-
-  const resetTaskImplementation = useCallback(
-    async (taskId: string): Promise<void> => {
-      const repoPath = requireActiveRepo(activeRepoPath);
-      try {
-        await host.taskResetImplementation(repoPath, taskId);
-        await Promise.all([
-          queryClient.invalidateQueries({
-            queryKey: agentSessionQueryKeys.list(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.qaReport(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.spec(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.plan(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-        ]);
-        await refreshTaskData(repoPath, taskId);
-        toast.success("Implementation reset", {
-          description: taskId,
-        });
-      } catch (error) {
-        toast.error("Failed to reset implementation", {
-          description: errorMessage(error),
-        });
-        throw error;
-      }
-    },
-    [activeRepoPath, queryClient, refreshTaskData],
-  );
-
-  const resetTask = useCallback(
-    async (taskId: string): Promise<void> => {
-      const repoPath = requireActiveRepo(activeRepoPath);
-      try {
-        await host.taskReset(repoPath, taskId);
-        await Promise.all([
-          queryClient.invalidateQueries({
-            queryKey: agentSessionQueryKeys.list(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.qaReport(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.spec(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-          queryClient.invalidateQueries({
-            queryKey: documentQueryKeys.plan(repoPath, taskId),
-            exact: true,
-            refetchType: "none",
-          }),
-        ]);
-        await refreshTaskData(repoPath, taskId);
-        toast.success("Task reset", {
-          description: taskId,
-        });
-      } catch (error) {
-        toast.error("Failed to reset task", {
-          description: errorMessage(error),
-        });
-        throw error;
-      }
-    },
-    [activeRepoPath, queryClient, refreshTaskData],
-  );
-
-  const transitionTask = useCallback(
-    async (taskId: string, status: TaskStatus, reason?: string): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.taskTransition(repoPath, taskId, status, reason);
-        },
-        successDescription: taskId,
-        failureTitle: "Failed to transition task",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const deferTask = useCallback(
-    async (taskId: string): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.taskDefer(repoPath, taskId, DEFERRED_BY_USER_REASON);
-        },
-        successTitle: "Task deferred",
-        successDescription: taskId,
-        failureTitle: "Failed to defer task",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const resumeDeferredTask = useCallback(
-    async (taskId: string): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.taskResumeDeferred(repoPath, taskId);
-        },
-        successTitle: "Task resumed",
-        successDescription: taskId,
-        failureTitle: "Failed to resume task",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const humanApproveTask = useCallback(
-    async (taskId: string): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.humanApprove(repoPath, taskId);
-        },
-        successTitle: "Task approved",
-        successDescription: taskId,
-        failureTitle: "Failed to approve task",
-      });
-    },
-    [runTaskMutation],
-  );
-
-  const humanRequestChangesTask = useCallback(
-    async (taskId: string, note?: string): Promise<void> => {
-      await runTaskMutation({
-        refreshStrategy: { kind: "task", taskId },
-        run: async (repoPath) => {
-          await host.humanRequestChanges(repoPath, taskId, note);
-        },
-        successTitle: "Changes requested",
-        successDescription: taskId,
-        failureTitle: "Failed to request changes",
-      });
-    },
-    [runTaskMutation],
-  );
-
   const clearTaskData = useCallback(() => {
-    manualRefreshTokenRef.current += 1;
-    setIsManualLoadingTasks(false);
-    lastTaskRefreshToastRef.current = null;
-    setDetectingPullRequestTaskId(null);
-    setLinkingMergedPullRequestTaskId(null);
-    setUnlinkingPullRequestTaskId(null);
-    setPendingMergedPullRequest(null);
-  }, []);
-
-  const tasks =
-    activeRepoPath && doneVisibleDays !== null ? (repoTaskDataQuery.data?.tasks ?? []) : [];
-  const isSettingsLoadingForActiveRepo = activeRepoPath !== null && settingsSnapshotQuery.isPending;
-  const isTaskQueryLoadingForActiveRepo =
-    activeRepoPath !== null && doneVisibleDays !== null && repoTaskDataQuery.isPending;
-  const isForegroundLoadingTasks =
-    isManualLoadingTasks || isSettingsLoadingForActiveRepo || isTaskQueryLoadingForActiveRepo;
-  const isRefreshingTasksInBackground =
-    activeRepoPath !== null && repoTaskDataQuery.isFetching && !isForegroundLoadingTasks;
-  const isLoadingTasks = isForegroundLoadingTasks;
+    taskReadFlow.clearTaskReadState();
+    pullRequestOperations.clearPullRequestState();
+  }, [pullRequestOperations.clearPullRequestState, taskReadFlow.clearTaskReadState]);
 
   return {
-    tasks,
-    isForegroundLoadingTasks,
-    isRefreshingTasksInBackground,
-    isLoadingTasks,
-    detectingPullRequestTaskId,
-    linkingMergedPullRequestTaskId,
-    unlinkingPullRequestTaskId,
-    pendingMergedPullRequest,
-    setIsLoadingTasks: setIsManualLoadingTasks,
+    tasks: taskReadFlow.tasks,
+    isForegroundLoadingTasks: taskReadFlow.isForegroundLoadingTasks,
+    isRefreshingTasksInBackground: taskReadFlow.isRefreshingTasksInBackground,
+    isLoadingTasks: taskReadFlow.isLoadingTasks,
+    detectingPullRequestTaskId: pullRequestOperations.detectingPullRequestTaskId,
+    linkingMergedPullRequestTaskId: pullRequestOperations.linkingMergedPullRequestTaskId,
+    unlinkingPullRequestTaskId: pullRequestOperations.unlinkingPullRequestTaskId,
+    pendingMergedPullRequest: pullRequestOperations.pendingMergedPullRequest,
+    setIsLoadingTasks: taskReadFlow.setIsLoadingTasks,
     clearTaskData,
-    refreshTaskData,
-    refreshTasksWithOptions,
-    refreshTasks,
-    syncPullRequests,
-    linkMergedPullRequest,
-    cancelLinkMergedPullRequest,
-    unlinkPullRequest,
-    createTask,
-    updateTask,
-    setTaskTargetBranch,
-    deleteTask,
-    resetTaskImplementation,
-    resetTask,
-    transitionTask,
-    deferTask,
-    resumeDeferredTask,
-    humanApproveTask,
-    humanRequestChangesTask,
+    refreshTaskData: taskReadFlow.refreshTaskData,
+    refreshTasksWithOptions: taskReadFlow.refreshTasksWithOptions,
+    refreshTasks: taskReadFlow.refreshTasks,
+    syncPullRequests: pullRequestOperations.syncPullRequests,
+    linkMergedPullRequest: pullRequestOperations.linkMergedPullRequest,
+    cancelLinkMergedPullRequest: pullRequestOperations.cancelLinkMergedPullRequest,
+    unlinkPullRequest: pullRequestOperations.unlinkPullRequest,
+    createTask: mutationCommands.createTask,
+    updateTask: mutationCommands.updateTask,
+    setTaskTargetBranch: mutationCommands.setTaskTargetBranch,
+    deleteTask: mutationCommands.deleteTask,
+    resetTaskImplementation: resetOperations.resetTaskImplementation,
+    resetTask: resetOperations.resetTask,
+    transitionTask: mutationCommands.transitionTask,
+    deferTask: mutationCommands.deferTask,
+    resumeDeferredTask: mutationCommands.resumeDeferredTask,
+    humanApproveTask: mutationCommands.humanApproveTask,
+    humanRequestChangesTask: mutationCommands.humanRequestChangesTask,
   };
 }
